+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added support for `Result.RowsAffected()` for `database/sql`
+
 ## v3.117.1
 * Fixed scan a column of type `Decimal(precision,scale)` into a struct field of type `types.Decimal{}` using `ScanStruct()`
 * Fixed race in integration test `TestTopicWriterLogMessagesWithoutData`